requirements.txt
pandas==1.0.3
python-dateutil==2.6.1
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
schedule==0.6.0
sendgrid==6.1.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
openpyxl==3.0.4
camelot-py[cv]==0.7.3
opencv-python-headless==4.1.2.30